% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBlock.R
\name{Find.block.distance}
\alias{Find.block.distance}
\title{Find the block by the snp distance and position}
\usage{
Find.block.distance(
  snplist,
  CHR,
  BP,
  CM = NULL,
  POS = NULL,
  distance_threshold = 5e+05
)
}
\description{
Find the block by the snp distance and position
}
