% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBlock.R
\name{Find.block.LD}
\alias{Find.block.LD}
\title{Find the block by block}
\usage{
Find.block.LD(snplist, CHR, BP, CM = NULL, Block.map)
}
\arguments{
\item{Block.map}{}
}
\description{
Find the block by block
}
