% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMLE.r
\name{getMLE}
\alias{getMLE}
\title{Calculate the weighted MLE}
\usage{
getMLE(x, y, w)
}
\arguments{
\item{x}{the covariate matrix of dimension n times p.}

\item{y}{a binary response vector of length n.}

\item{w}{a weight vector of length n.}
}
\description{
This function calculates the MLE for a weighted logistic regression.
}
