% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxDetect.R
\name{MinDetect}
\alias{MinDetect}
\title{Detect function}
\usage{
MinDetect(P1, P2, K = 200, F.null = NULL)
}
\arguments{
\item{P1}{}

\item{P2}{}

\item{K}{}

\item{F.null}{}
}
\description{
Detect function
}
