% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxID.R
\name{MinID}
\alias{MinID}
\title{Identification of sharing association}
\usage{
MinID(P1, P2, K = 200, F.null = NULL, type, alpha)
}
\arguments{
\item{F.null}{null distribution}

\item{type}{Criteria for the identification}

\item{alpha}{error rate level}
}
\value{

}
\description{
Identification of sharing association
}
