% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMaxP.R
\name{MinMaxP.geno}
\alias{MinMaxP.geno}
\title{Detect function for the genotype data}
\usage{
MinMaxP.geno(
  P1,
  P2,
  ref.geno = NULL,
  ref.bed = NULL,
  block.map = NULL,
  var.name = NULL,
  output.dir = "./temp/",
  block.thresh = 0.995,
  K = 1000,
  exact = F
)
}
\arguments{
\item{F.null}{}
}
\value{

}
\description{
Detect function for the genotype data
}
\examples{



}
