% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMaxP_simu.R
\name{MinMaxP.simu}
\alias{MinMaxP.simu}
\title{general detection function}
\usage{
MinMaxP.simu(P1, P2, var.name = NULL, block.map = NULL, corr = NULL)
}
\arguments{
\item{P2}{}
}
\description{
general detection function
}
