% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meff.R
\name{N_eff}
\alias{N_eff}
\title{calculate the effective number from a genotype file}
\usage{
N_eff(
  var.name.target,
  ref.geno = NULL,
  ref.bed = NULL,
  block.map = NULL,
  var.name = NULL,
  output.dir = "./temp/",
  block.thresh = 0.995,
  K = 1000,
  exact = F
)
}
\description{
calculate the effective number from a genotype file
}
