% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAT.estimation.r
\name{SAT.estimation}
\alias{SAT.estimation}
\title{SAT estimation based on pooled subsample}
\usage{
SAT.estimation(S, X, beta.pilot, stage1.index, stage2.index, stage1.weights,
               stage1.y, stage2.y, method = "SAT-S")
}
\arguments{
\item{S}{a binary vector of length n. Surrogate observations for all samples.}

\item{X}{a matrix of dimension n times p (the first column needs to be 1). The covariate matrix contains observations for all n samples.}

\item{beta.pilot}{the pilot estimator.}

\item{stage1.index}{a vector of length r1. The index of pilot sampled patients.}

\item{stage2.index}{a vector of length r. The index of second-stage sampled patients.}

\item{stage1.weights}{a vector of weights for patients who are selected in pilot sampling.}

\item{stage1.y}{a binary vector of length r1. The manual chart review results for patients in \code{stage1.index}.}

\item{stage2.y}{a binary vector of length r. The manual chart review results for patients in \code{stage2.index}.}

\item{method}{two methods are available: SAT-S or SAT-cY.}
}
\value{
The function returns the final SAT estimates of the association coefficients.
}
\description{
This function implements the final SAT estimation based on data pooled from two stages of sampling.
A weighted logistic regression is conducted.
}
\examples{
library(SAT)
set.seed(0)

colnames(lung_cancer)
X <- cbind(1, lung_cancer[,3:5])
Y <- lung_cancer[,1]
S <- lung_cancer[,2]

# pilot sampling
stage1.index <- SAT.stage1.sampling(r1 = 400, n = 1e5, S, Rpar = 0.5)
# true phenotype collection
stage1.y <- Y[stage1.index]

# second stage sampling
stage2 <- SAT.stage2.sampling(r1 = 400, n = 1e5, S, Rpar = 0.5, r = 800,
                              stage1.index, stage1.y, X, method = "SAT-cY")
# true phenotype collection
stage2.y <-  Y[stage2$stage2.index]

# final estimation
SAT.est <- SAT.estimation(S, X, beta.pilot = stage2$beta.pilot, stage1.index = stage1.index,
               stage2.index = stage2$stage2.index,
               stage1.weights = stage2$stage1.weights,
               stage1.y = stage1.y, stage2.y = stage2.y,
               method = "SAT-cY")
SAT.est$SAT.estimate

}
\references{
Liu, X., Chubak, J., Hubbard, R. A. & Chen, Y. (2021).
SAT: a Surrogate Assisted Two-wave case boosting sampling method, with application to EHR-based association studies.
}
