% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMLE.r
\name{getMLE}
\alias{getMLE}
\title{Calculate the weighted MLE}
\usage{
getMLE(x, y, w)
}
\arguments{
\item{x}{the input covariate matrix}

\item{y}{the input response vector}

\item{w}{the weight vector}
}
\description{
This function calculates the MLE for a weighted logistic regression.
}
