% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxID.R
\name{MinID}
\alias{MinID}
\title{Identification of sharing association}
\usage{
MinID(
  P1,
  P2,
  type = "FWER",
  alpha = 0.05,
  eff.ratio = 0.25,
  ref.geno = NULL,
  ref.bed = NULL,
  block.map = NULL,
  var.name = NULL,
  output.dir = "./temp/",
  simple = T,
  K = 1000
)
}
\arguments{
\item{P1}{the input is p-value sequence}

\item{type}{Criteria for the identification}

\item{alpha}{error rate level}

\item{F.null}{null distribution}
}
\value{
return is S and tau, where tau is the p-value threshold
}
\description{
Identification of sharing association
}
\examples{



}
