% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMaxP.R
\name{MinMaxP.detect}
\alias{MinMaxP.detect}
\title{Detect function for the genotype data}
\usage{
MinMaxP.detect(P.max, P.min, M)
}
\arguments{
\item{M}{}
}
\value{
detection pvalue
}
\description{
Detect function for the genotype data
}
