% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMaxP.R
\name{MinMaxP.discov}
\alias{MinMaxP.discov}
\title{Detect function for the genotype data}
\usage{
MinMaxP.discov(P.max, P.min, method = c("FWER", "FDR"), alpha = 0.05)
}
\arguments{
\item{method}{FWER and FDR}

\item{alpha}{significance value}
}
\value{
detection pvalue
}
\description{
Detect function for the genotype data
}
