% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meff.R
\name{N_eff}
\alias{N_eff}
\title{calculate the effective number of given locations}
\usage{
N_eff(
  block.list,
  ref.geno = NULL,
  block.thresh = 0.995,
  ind.row = rows_along(ref.geno)
)
}
\arguments{
\item{block.map}{}
}
\description{
calculate the effective number of given locations
}
